fix(ffi): wrong extern-fn return types read undefined register bits
getBeepVolume was declared () i32 in frontier_addons (real symbol is
uint16_t) — the i32 return reads undefined high bits of eax for a value
that is then used to save/restore beep volume. nimBufferToDisplayBuffer
was declared to return [*c]u8 while the real ported symbol returns void,
so 7 call sites materialized an uninitialized register as a pointer
before discarding it. Also normalized the lone (u16,u16,u16)
displayCalcErrorMessage outlier in the matrix-editor-entry owner to the
canonical (u8, i16, i16) shared by ~65 other declarations.

+++ zig_src/frontier/frontier_matrix_editor_entry_owned.zig
@@ -86,4 +86,4 @@ extern fn setJRegisterAsInt(as_array_pointer: bool, to_store: i16) void;
 extern fn showMatrixEditor() void;
 extern fn refreshScreen(source: u16) void;
 extern fn printTraceMatElement(where: u16) void;
-extern fn displayCalcErrorMessage(error_code: u16, line1: u16, line2: u16) void;
+extern fn displayCalcErrorMessage(error_code: u8, line1: i16, line2: i16) void;
